+++ app/Filters/JobFilter.php
@@ -3,6 +3,24 @@
 class JobFilter
 {
     /**
+     * Returns an array of jobs each with only the attributes specified
+     *
+     * @param array $jobs
+     * @param array $fields
+     *
+     * @return array
+     */
+    public function filterFields($jobs = [], $fields = [])
+    {
+        foreach ($jobs as &$job) {
+            $job = array_filter($job, function ($key) use ($fields) {
+                return in_array($key, $fields);
+            }, ARRAY_FILTER_USE_KEY);
+        }
+        return $jobs;
+    }
+
+    /**
      * Sort jobs by date posted, desc
      *
      * @param array $jobs